THUCYDIDES-160: add ClassFinder tests for classes located in JAR files
When they are run from an IDE like IntelliJ, the tests that use a maven
dependency do not test finding classes that are packaged in a JAR file.

This is because IntelliJ will setup the dependencies so that one module
depends on another module's .class files directly. This way, the IDE
does not have to waste time to package classes to jar files, and the
development is faster.