THUCYDIDES-160: fix issue
This commit fixes the loading of classes from jar files or file-system
directories, when the path contains characters which need to be
URL-encoded (e.g. the space character).

The existing code was taking only the space character into
consideration, whereas this commit fixes the issue in all situations:
every time the path contains characters that need to be URL-encoded.